% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Natl_mvmt.R
\name{Natl_mvmt}
\alias{Natl_mvmt}
\title{Natl_mvmt}
\usage{
Natl_mvmt(x, y, z)
}
\arguments{
\item{x}{year of last election}

\item{y}{year of current election}

\item{z}{Is the candidate a Republican}
}
\value{
The movement in national congressional popular vote
}
\description{
returns the change in national performance
}
\details{
Details
}
\author{
Matthew Brown
}
