% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Results_Predictor.R
\name{Results_Predictor}
\alias{Results_Predictor}
\title{Results_Predictor}
\usage{
Results_Predictor(x, w, z, b, c, d)
}
\arguments{
\item{x}{year of last election}

\item{w}{The Political Party of candidate a (1 = Republican, 2 = Democrat)}

\item{z}{Previous Margin}

\item{b}{Predicted National Margin at time of election}

\item{c}{The amount of money candidate a is expected to raise}

\item{d}{The amount of money candidate b is expected to raise}
}
\value{
The predicted margin of loss or victory for candidate a.
}
\description{
Calculates the predicted margin of loss or victory for a candidate.
}
\author{
Matthew Brown
}
