% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/YPVI_Finder.R
\name{YPVI_Finder}
\alias{YPVI_Finder}
\title{YPVI_Finder}
\usage{
YPVI_Finder(x)
}
\arguments{
\item{x}{year of election}
}
\value{
The movement in national congressional popular vote
}
\description{
returns the change in national performance
}
\details{
Details
}
\examples{

}
\author{
Matthew Brown
}
