% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fundraising_Predictor.R
\name{Fundraising_Predictor}
\alias{Fundraising_Predictor}
\title{yp
Fundraising_Predictor}
\usage{
Fundraising_Predictor(x, y, z, b, c)
}
\arguments{
\item{x}{year of last election}

\item{y}{The Political Party of candidate a (1 = Republican, 2 = Democrat)}

\item{z}{Previous Margin}

\item{b}{Predicted National Margin at time of election}

\item{c}{The amount of money candidate b is expected to raise}
}
\value{
The amount of money a candidate should raise in order to be competative
}
\description{
returns the yearly performance table
}
\details{
Details
}
\author{
Matthew Brown
}
